requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-socketio
flask-jwt-extended
PyJWT
pymysql
python-dotenv
requests
bcrypt
pydantic
redis
eventlet
cryptography
